feat(dashboard): make dashboard padding togglable
Hardcoded padding top causes extensions made with this SDK looks strange
and hence we will disable it by default for new extension sdk. we figure
that out by sending the sdkType with init call we could differentiate
the SDKs.

Secondly, we have added two methods in frame module to control the
padding.

fixes: CS-29720