feat: add branches support
We have added two methods to get current and all the branches. I have
written the test cases to test this behaviour when the branch is present
and when the branch is absent.

Moreover, I have added types for the stack objects and accordingly
refactored the enire codebase to use it.

fixes: https://contentstack.atlassian.net/browse/CS-32161