update: move org_uid to headers
To ensure that the getAllStacks only receive the list of stacks that the
user is currently a member of, we updated the url to /stacks. Now, this
route accepts the organization uid through headers.

Also, this route now uses getStacks instead of getAllStacks as that
route was redundant. But, the event adds the api key by default. So, we
have added another key to skip the api key in the headers.